Update nfs-ha-vm.json
Trying to add timeout to the right place.  We are timing out on the setup_nfs_ha deployment for one of the VMs.. maybe the load balancer... but I think it's the VM because we have a large amount of storage getting allocated.  So we need to extend the timeout of the VM. MIght be 90 minutes default for east coast... could try different region.